Renames zipfile_parser -> archive_parser
